refactor(alertd): share the check age formatter
The coarse seconds-to-single-unit formatter fhir_jobs used for its summaries
is needed by another check, so move it to the shared check helpers rather than
add a fourth copy alongside the ones in external_users and http_errors.
